% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors.gam}
\alias{calcParaVectors.gam}
\title{A calcParaVectors Method for gam (mgcv package) Objects}
\usage{
\method{calcParaVectors}{gam}(fit, family_use, new_covariate, total_cells, ...)
}
\arguments{
\item{fit}{a fitted object in the marginal_list.}

\item{family_use}{a string scalar or vector of marginal distribution used.}

\item{new_covariate}{a cell-by-covariate data frame obtained in the list output from
\code{\link{constructDataPop}}. It must have a corr_group variable.}

\item{total_cells}{a positive integer for the number of total cells to simulate.}

\item{...}{additional arguments passed to calcParaVectors S3 method functions.}
}
\description{
A calcParaVectors Method for gam (mgcv package) Objects
}
