% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors.gamlss}
\alias{calcParaVectors.gamlss}
\title{A calcParaVectors Method for gamlss Objects}
\usage{
\method{calcParaVectors}{gamlss}(fit, family_use, new_covariate, data, ...)
}
\arguments{
\item{fit}{a fitted object in the marginal_list.}

\item{family_use}{a string scalar or vector of marginal distribution used.}

\item{new_covariate}{a cell-by-covariate data frame obtained in the list output from
\code{\link[=constructDataPop]{constructDataPop()}}. It must have a corr_group variable.}

\item{data}{a cell-by-covariate data frame used to fit marginal models in
\code{\link[=fitMarginalPop]{fitMarginalPop()}}. It must have a corr_group variable.}

\item{...}{additional arguments passed to calcParaVectors S3 method functions.}
}
\description{
A calcParaVectors Method for gamlss Objects
}
