% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors.glmmTMB}
\alias{calcParaVectors.glmmTMB}
\title{A calcParaVectors Method for glmmTMB Objects}
\usage{
\method{calcParaVectors}{glmmTMB}(fit, family_use, new_covariate, ...)
}
\arguments{
\item{fit}{a fitted object in the marginal_list.}

\item{family_use}{a string scalar or vector of marginal distribution used.}

\item{new_covariate}{a cell-by-covariate data frame obtained in the list output from
\code{\link[=constructDataPop]{constructDataPop()}}. It must have a corr_group variable.}

\item{...}{additional arguments passed to calcParaVectors S3 method functions.}
}
\description{
A calcParaVectors Method for glmmTMB Objects
}
