% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkVectorContain}
\alias{checkVectorContain}
\title{Check membership of first vector compared to other vectors}
\usage{
checkVectorContain(...)
}
\arguments{
\item{...}{arguments of two or more vectors}
}
\value{
a logical scalar
}
\description{
Check membership of first vector compared to other vectors
}
\examples{
NULL
}
