% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkVectorEqual}
\alias{checkVectorEqual}
\title{Check if multiple vectors have same elements}
\usage{
checkVectorEqual(..., ignore_order = TRUE)
}
\arguments{
\item{...}{arguments of two or more vectors}

\item{ignore_order}{logical scalar to disregard the ordering of elements in
input vectors. Default is TRUE.}
}
\value{
a logical scalar
}
\description{
This is an internal helper function to check if two or more vectors have same
elements (with their order considered as an option)
}
\examples{
vec1 <- c("cherry", "apple", "banana", "cherry")
vec2 <- c("cherry", "apple", "cherry", "banana")
vec3 <- c("banana", "cherry", "apple", "cherry")

checkVectorEqual(vec1, vec2, vec3, ignore_order = TRUE)  # returns TRUE
checkVectorEqual(vec1, vec2, vec3, ignore_order = FALSE)  # returns FALSE
}
