% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructDataPop.R
\name{constructDataPop}
\alias{constructDataPop}
\title{Construct a list of input data}
\usage{
constructDataPop(
  sce,
  eqtlgeno_df,
  new_eqtlgeno_df = NULL,
  new_covariate = NULL,
  overlap_features = NULL,
  sampid_vec = NULL,
  ct_copula = TRUE,
  slot_name = "counts",
  snp_model = c("single", "multi"),
  cellstate_colname = "cell_type",
  feature_colname = "gene_id",
  snp_colname = "snp_id",
  loc_colname = "POS",
  chrom_colname = "CHR",
  indiv_colname = "indiv",
  prune_thres = 0.9
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{eqtlgeno_df}{a dataframe with eQTL annotations and SNP genotypes for each gene}

\item{new_eqtlgeno_df}{a dataframe with eQTL annotations and SNP genotypes for each
gene in new individuals}

\item{new_covariate}{a cell covariate dataframe for which to simulate}

\item{overlap_features}{an optional string vector to filter for features}

\item{sampid_vec}{an optional string vector to filter for sample ids}

\item{ct_copula}{a logical scalar for whether to fit the copula by cell type}

\item{slot_name}{a string scalar to specify the input SCE slot used}

\item{snp_model}{a string scalar to specify the type of SNP model used}

\item{cellstate_colname}{a string scalar for the cell state variable in eqtlgeno_df
and cell covariate of SCE}

\item{feature_colname}{a string scalar for the feature variable in eqtlgeno_df}

\item{snp_colname}{a string scalar for the SNP variable in eqtlgeno_df}

\item{loc_colname}{a string scalar for the last column of eQTL annotation in eqtlgeno_df}

\item{chrom_colname}{a string scalar for the chromosome variable in eqtlgeno_df}

\item{indiv_colname}{a string scalar for the sample ID variable in cell covariate of SCE}

\item{prune_thres}{numerical value between 0 and 1 to threshold the SNPs}
}
\value{
outputs a list with following elements:
\describe{
\item{\code{count_mat}}{a cell-by-gene matrix of response values.}
\item{\code{covariate}}{a cell-by-covariate data frame used for fit marginal.}
\item{\code{new_covariate}}{an optional cell-by-covariate data frame used for prediction.}
\item{\code{important_features}}{a string vector of gene ids.}
\item{\code{eqtl_geno_list}}{a list of eQTL genotype dataframes for each gene
(for fit marginal).}
\item{\code{new_eqtl_geno_list}}{a optional list of eQTL genotype dataframes
for each gene (for prediction on new individuals).}
\item{\code{filtered_gene}}{string vector of features QC filtered.}
}
}
\description{
Function extracts an expression matrix, cell covariates, and filters the SNPs
in the eQTL genotype dataframe.
}
\examples{
NULL
}
