% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructEqtlGeno.R
\name{constructEqtlGeno}
\alias{constructEqtlGeno}
\title{Construct eQTL annotation and genotype dataframe}
\usage{
constructEqtlGeno(
  eqtl_annot_df,
  geno_df,
  sampid_vec,
  name = NULL,
  feature_colname = "gene_name",
  celltype_colname = "cell_type",
  snp_colname = "snp_id",
  loc_colname = "POS"
)
}
\arguments{
\item{eqtl_annot_df}{eQTL annotation dataframe}

\item{geno_df}{eQTL by sample genotype dataframe}

\item{sampid_vec}{sample id vector}

\item{name}{string or integer scalar to identify input}

\item{feature_colname}{feature variable name}

\item{celltype_colname}{cell type variable name}

\item{snp_colname}{SNP id variable name}

\item{loc_colname}{SNP position variable name}
}
\value{
a dataframe with eQTL annotations and genotype of SNPs
}
\description{
This function creates a combined dataframe using input eQTL annotations, and sample genotype dataframe.
}
\examples{
NULL
}
