% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMarginalPop.R
\name{constructFormula}
\alias{constructFormula}
\title{Constructs a Model Formula}
\usage{
constructFormula(
  model_formula,
  interact_colnames = NULL,
  snp_cov = NULL,
  force_formula = FALSE
)
}
\arguments{
\item{model_formula}{a string scalar to specify the model formula, including
random effects (if any) but without SNP genotypes or SNP genotype interaction effects.}

\item{interact_colnames}{a string scalar or vector for the variable names that
have first-order interaction with SNP genotypes.}

\item{snp_cov}{a string scalar or vector that specifies the SNP covariates used in the model.}

\item{force_formula}{a logical scalar for whether to bypass model parsimony check.
If \code{force_formula = TRUE}, interaction terms whose covariates are not
main effects in the model would be permitted. Results in error if
\code{force_formula = FALSE} and \code{length(geno_interact_names) > 0}.}
}
\value{
a formula object
}
\description{
A helper function that specifies the formula for the marginal model.
}
\examples{
NULL
}
