% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{constructPAFormula}
\alias{constructPAFormula}
\title{Construct a model formula for power analysis}
\usage{
constructPAFormula(
  fm,
  method = c("nb", "poisson", "gaussian", "pseudoBulkLinear"),
  snpid = NULL,
  cellstate_colname = NULL
)
}
\arguments{
\item{fm}{a stats::formula object from the full marginal model.}

\item{method}{a character object specifying the method that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{snpid}{a character object contains snpid.}

\item{cellstate_colname}{a string scalar specifying the cell state variable
The default is "cell_type".}
}
\value{
a stats::formula object for power analysis in each specified type.
}
\description{
Construct a model formula for power analysis
}
\examples{
NULL
}
