% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSCE.R
\name{constructSCE}
\alias{constructSCE}
\title{Construct a SingleCellExperiment object with specified covariates.}
\usage{
constructSCE(
  data_obj,
  cellcov_df = NULL,
  featcov_df = NULL,
  cellcov_names,
  sampid_vec,
  overlap_features = NULL,
  assay_name = "RNA",
  slot_name = "counts",
  sce_name = slot_name,
  indiv_colname = "indiv",
  cellstate_colname = "cell_type",
  factor_colnames = NULL,
  cellcov_renames = cellcov_names
)
}
\arguments{
\item{data_obj}{Seurat, sce, or expression matrix (gene-by-cell) used to construct the new sce object}

\item{cellcov_df}{dataframe containing covariates for all cells. Row names must have cell names that match data_obj. Only used when data_obj is a matrix object.}

\item{featcov_df}{dataframe containing covariates for all features (ie genes). Row names must have feature names that match data_obj. Only used when data_obj is a matrix object.}

\item{cellcov_names}{string vector of all columns to extract from cell covariates}

\item{sampid_vec}{string vector of sample ids (individuals)}

\item{overlap_features}{string vector for features to filter data_obj. All features must be present in the data_obj}

\item{assay_name}{name of the Seurat object slot of interest. Only used if data_obj is Seurat object}

\item{slot_name}{type of assay data (either logcounts or counts). Default is 'counts'. Only used if data_obj is Seurat or sce object}

\item{sce_name}{type of assay data for output sce}

\item{indiv_colname}{string scalar for the sample id colname. Must be present in cellcov_df or data_obj. Default is 'indiv'}

\item{cellstate_colname}{string scalar for the cell state colname. Must be present in cellcov_df or data_obj. Default is 'cell_type'}

\item{factor_colnames}{string vector or scalar for the columns to coerce into factor variables. Must be present in cellcov_df or data_obj. Default is NULL}

\item{cellcov_renames}{optional string vector to rename the cell covariates. Must be same length as cellcov_names. Default is same as cellcov_names}
}
\value{
a SingleCellExperiment object
}
\description{
Construct a SingleCellExperiment object with specified covariates.
}
\examples{
NULL
}
