% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPbulkGeno.R
\name{createPbulkExprGeno}
\alias{createPbulkExprGeno}
\title{Creates pseudobulk expression grouped by eQTL genotypes.}
\usage{
createPbulkExprGeno(
  sce_list,
  eqtlgeno,
  feature_sel,
  celltype_sel = "allct",
  eqtl_indx = 1L,
  eqtl_snp = NULL,
  normalize_type = c("log1p", "seurat", "scran", "none"),
  aggregate_type = c("mean", "sum"),
  slot_name = c("counts", "logcounts"),
  rescale = FALSE,
  overwrite = FALSE,
  feature_colname = "gene_id",
  snp_colname = "snp_id",
  indiv_colname = "indiv",
  celltype_colname = "cell_type",
  loc_colname = "POS",
  if_plot = TRUE,
  ...
)
}
\arguments{
\item{sce_list}{A list of named SingleCellExperiment (SCE) objects. Each SCE
object must have the \code{indiv_colname} and \code{celltype_colname}
columns in its colData. If no names are provided in the list, the objects
will be named "sce_1", "sce_2", etc, by default.}

\item{eqtlgeno}{A dataframe of eQTL with genotypes for samples. Must have the
columns corresponding to \code{feature_colname}, \code{snp_colname},
\code{celltype_colname}, and \code{loc_colname}. The \code{loc_colname}
variable must be the last column that precedes the genotypes of the
individuals (samples).}

\item{feature_sel}{A string scalar for the feature name (eg. a gene id) to
select. Must exist in rowname of every SCE object in \code{sce_list}.}

\item{celltype_sel}{A string scalar or vector of the cell states (eg. cell types)
to select. Must be one of the cell types in the \code{celltype_colname} of
every SCE object in \code{sce_list}.}

\item{eqtl_indx}{An integer scalar or vector to select the eQTL from \code{eqtlgeno}
for a given celltype and feature. Only used if \code{eqtl_snp = NULL}. The default is 1.
The values correspond to \code{celltype_sel} and are recycled if the length of
elements is shorter.}

\item{eqtl_snp}{A string scalar or vector to select the eQTL from \code{eqtlgeno}.
Overrides \code{eqtl_indx} option if both \code{eqtl_indx} and \code{eqtl_snp}
are specified. Can select any SNP in the \code{eqtlgeno} (even from another
feature). Must exist in the \code{snp_colname} column of \code{eqtlgeno}.
The values correspond to \code{celltype_sel} and are recycled if the length of
elements is shorter.}

\item{normalize_type}{A string scalar for normalization method applied to
\code{sce_list}.}

\item{aggregate_type}{A string scalar for pseudobulk aggregation applied to
\code{sce_list} after subsetting by \code{celltype_sel}.}

\item{slot_name}{A string scalar for the SCE slot used. The default is "counts".}

\item{rescale}{A logical scalar for whether to rescale the response values
based on maximum expression value.}

\item{overwrite}{A logical value on whether to overwrite the "logcounts" slot in
input sce object. Must be set to \code{overwrite = TRUE} to overwrite. The
default is FALSE.}

\item{feature_colname}{A string scalar for column name of feature variable
(i.e. gene). The default is "gene_id".}

\item{snp_colname}{A string scalar for column name of the SNP variable. The
default is "snp_id".}

\item{indiv_colname}{A string scalar for column name of individuals (samples).
The default is "indiv".}

\item{celltype_colname}{A string scalar for column name of cell type. The
default is "cell_type".}

\item{loc_colname}{A string scalar for column name of SNP position variable.
The default is "POS".}

\item{if_plot}{A logical scalar specifying whether to return a pseudobulk plot.
The default is TRUE.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Outputs a dataframe with following columns:
\describe{
\item{\code{indiv}}{Name of individual's sample ID.}
\item{\code{response}}{Pseudobulk expression of given feature.}
\item{\code{genotype}}{Genotype values of individual for selected eQTL SNP.}
\item{\code{feature}}{Name of selected feature.}
\item{\code{celltype}}{Name of selected cell type.}
\item{\code{snp}}{Name of selected eQTL SNP.}
\item{\code{sce_name}}{Name of the SingleCellExperiment object.}
\item{\code{normalization}}{Type of normalization method used.}
\item{\code{aggregate_type}}{Type of pseudobulk aggregation used.}
\item{\code{slot_used}}{Slot used from input SingleCellExperiment object.}
}
}
\description{
Computes pseudobulk based on normalization and aggregation options, and combines
with eQTL genotypes to output dataframe of pseudobulk expression vs genotypes.
}
\examples{
NULL
}
