% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_eqtlgeno_Bcell}
\alias{example_eqtlgeno_Bcell}
\title{Example B cell eQTL genotype data}
\format{
A tibble with 2,406 rows and 106 columns:
\describe{
\item{\code{cell_type}}{Cell type (e.g., \code{"cd4nc"}, \code{"cd8nc"}, \code{"nk"}, \code{"cd4et"}, etc.).}
\item{\code{gene_name}}{Gene symbol (e.g., \code{"PLEKHO1"}, \code{"TNFRSF1B"}, \code{"UTS2"}).}
\item{\code{gene_id}}{Ensembl gene ID.}
\item{\code{snp_id}}{SNP identifier in \code{CHR:POS} format.}
\item{\code{CHR}}{Chromosome number.}
\item{\code{POS}}{Genomic position (base-pair coordinate).}
\item{\code{SAMP1}, \code{SAMP2}, \dots}{Genotype dosage (0/1/2) for each
individual. The remaining columns \code{SAMPk} store genotypes for all
samples used in the example.}
}
}
\usage{
data("example_eqtlgeno_Bcell")
}
\description{
A tibble containing example genotype data for selected cis-eQTL SNPs
for the selected B cells from OneK1K. This
dataset can be used together with \code{example_sce_Bcell} to show
dynamic or cell-type–specific eQTL modeling.
No eQTL effect size results are included. Genotypes are all permuted.
}
\examples{
data("example_eqtlgeno_Bcell")
example_eqtlgeno_Bcell
dplyr::count(example_eqtlgeno_Bcell, gene_name)
}
\keyword{datasets}
