% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSCE.R
\name{extractFromSCE}
\alias{extractFromSCE}
\title{Extract data from SingleCellExperiment object}
\usage{
extractFromSCE(
  sce_obj,
  features = NULL,
  slot_name = "counts",
  indiv_colname = "indiv",
  ...
)
}
\arguments{
\item{sce_obj}{a SingleCellExperiment object}

\item{features}{string vector to filter for features}

\item{slot_name}{string scalar to specify the type of input assay data}

\item{indiv_colname}{string scalar to specify the column that has the indiv ids}

\item{...}{other options to specify extracted outputs}
}
\value{
a list of extracted data
}
\description{
Extract data from SingleCellExperiment object
}
\examples{
NULL
}
