% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSCE.R
\name{extractFromSeurat}
\alias{extractFromSeurat}
\title{Extract data from Seurat object}
\usage{
extractFromSeurat(
  seurat_obj,
  features = NULL,
  assay_name = "RNA",
  slot_name = "counts",
  indiv_colname = "indiv",
  ...
)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{features}{string vector to filter for features}

\item{assay_name}{string scalar to specify the name of input assay data}

\item{slot_name}{string scalar to specify the type of input assay data}

\item{indiv_colname}{string scalar to specify the column that has the indiv ids}

\item{...}{other options to specify extracted outputs. Currently available:
\itemize{
\item \code{col_class} : logical value
\item \code{feat_names} : logical value
\item \code{cell_names} : logical value
\item \code{sc_indiv} : logical value
\item \code{expr_mat} : logical value
\item \code{sc_cov} : logical value
\item \code{feat_cov} : logical value
}}
}
\value{
a list of extracted data with following optional outputs:
\itemize{
\item \code{col_class} : a vector of the column classes for cell covariate variables
\item \code{feat_names} : a vector of feature names
\item \code{cell_names} : a vector of cell names
\item \code{sc_indiv} : a vector of distinct sample ids extracted from cell covariate using \code{indiv_colname}
\item \code{expr_mat} : an expression matrix extracted using \code{assay_name} and \code{slot_name} options
\item \code{sc_cov} : a dataframe for cell covariates
\item \code{feat_cov} : a dataframe for feature covariates
}
}
\description{
Extract data from Seurat object
}
\examples{
NULL
}
