% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCopulaPop.R
\name{fitCopulaPop}
\alias{fitCopulaPop}
\title{Fits copula for input}
\usage{
fitCopulaPop(
  sce,
  assay_use,
  input_data,
  marginal_list,
  family_use,
  copula = "gaussian",
  DT = TRUE,
  pseudo_obs = FALSE,
  epsilon = 1e-06,
  family_set = c("gaussian", "indep"),
  important_feature = "all",
  n_cores,
  parallelization = "mcmapply",
  BPPARAM = NULL
)
}
\arguments{
\item{sce}{a SingleCellExperiment object.}

\item{assay_use}{add later}

\item{input_data}{add later}

\item{marginal_list}{add later}

\item{family_use}{a string scalar to specify model fitting used.}

\item{copula}{a string value to specify the type of Copula fitting to use.
Currently, only Gaussian copula is supported. The default is "gaussian".}

\item{DT}{add later}

\item{pseudo_obs}{add later}

\item{epsilon}{a numeric value close to 0 to specify tolerance for avoiding
0 or 1 quantiles. The default value is 1e-06.}

\item{family_set}{add later}

\item{important_feature}{add later}

\item{n_cores}{positive integer value (greater or equal to 1) to specify the
number of CPU threads used in parallelization.}

\item{parallelization}{a string scalar specifying the type of parallelization
used during marginal fitting.}

\item{BPPARAM}{add later}
}
\value{
outputs a list with following elements:
}
\description{
Fits copula for input
}
\examples{
NULL
}
