% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMarginalPop.R
\name{fitMarginalPop}
\alias{fitMarginalPop}
\title{Fit marginal models for every feature}
\usage{
fitMarginalPop(
  data_list,
  mean_formula,
  model_family = "nb",
  interact_colnames = NULL,
  parallelization = c("pbmcapply", "future.apply", "parallel", "biocparallel"),
  n_cores = 2L,
  loc_colname = "POS",
  snp_colname = "snp_id",
  celltype_colname = "cell_type",
  indiv_colname = "indiv",
  filter_snps = TRUE,
  snpvar_thres = 0,
  force_formula = FALSE,
  keep_cellnames = FALSE,
  BPPARAM = NULL,
  future.seed = FALSE,
  data_maxsize = 1,
  ...
)
}
\arguments{
\item{data_list}{a list of input data.}

\item{mean_formula}{a string scalar to specify the mean formula, including
random effects (if any) but without SNP genotypes or SNP genotype interaction effects.}

\item{model_family}{a string scalar to specify model fitting used.}

\item{interact_colnames}{a string scalar or vector for the variable names that
have first-order interaction with SNP genotypes.}

\item{parallelization}{a string scalar specifying the parallelization backend
used during marginal fitting. Must be one of either "parallel", "future.apply",
"biocparallel", or "pbmcapply". The default value is "pbmcapply". See details.}

\item{n_cores}{positive integer value (greater or equal to 1) to specify the
number of CPU cores used in parallelization. The default is 2.}

\item{loc_colname}{a string scalar for column name of SNP position variable.}

\item{snp_colname}{a string scalar for column name of SNP id variable.}

\item{celltype_colname}{a string scalar for column name of cell type.}

\item{indiv_colname}{a string scalar for column name of individuals (samples).}

\item{filter_snps}{a logical scalar for whether to filter out SNP covariates
with either low-variance or with only 1 distinct genotype (ie. all 1's)
prior to fitting the model.}

\item{snpvar_thres}{a numeric scalar (between 0 to 1) used to filter out SNPs
whose variance of genotypes across samples are below this threshold.
Used together when \code{filter_snps = TRUE}.}

\item{force_formula}{a logical scalar for whether to bypass model parsimony check.
If \code{force_formula = TRUE}, interaction terms whose covariates are not
main effects in the model would be permitted. Results in error if
\code{force_formula = FALSE} and \code{length(geno_interact_names) > 0}.}

\item{keep_cellnames}{a logical scalar for whether to keep cell barcode names.
If \code{keep_cellnames = TRUE}, the memory will be larger. The default is FALSE.}

\item{BPPARAM}{a BiocParallelParam class object (from \code{BiocParallel} R package)
that must be specified when using \code{parallelization = "biocparallel"}. Either
\code{BiocParallel::SnowParam()} or \code{BiocParallel::MulticoreParam()}
can be used to initialize, depending on the operating system. BPPARAM is
not used in other parallelization options. The default is NULL.}

\item{future.seed}{a logical or an integer (of length one or seven), or a list
of length(X) with pre-generated random seeds that can be specified when using
\code{parallelization = "future.apply"}. See \code{future.apply::future_eapply}
documentation for more details on its usage. future.seed is not used in
other parallelization options. The default is FALSE.}

\item{data_maxsize}{a positive numeric value used to set max data_list size
in GiB increments. Used only when \code{parallelization = "future.apply"}.
The default is 1.}

\item{...}{additional arguments passed to internal functions.}
}
\value{
a list of named features, each containing a list with the following items:
\describe{
\item{\code{fit}}{a \code{glmmTMB} fit object.}
\item{\code{time}}{a numeric scalar for the elapsed time to fit the given feature.}
\item{\code{snp_cov}}{a string scalar or vector of SNP ids used in the
fit for given feature.}
\item{\code{model_attr}}{a list of attributes extracted from each model. NOT
currently implemented.}
\item{\code{removed_cell}}{a string scalar or vector of the cell names
removed due to low-variance (NOT currently implemented).}
}
}
\description{
Fits a specified parametric model using various input parameters.
}
\details{
\subsection{Parallelization options}{

If "parallel" is used then \code{mclapply} is called from the \code{parallel} package;
if "biocparallel" is used, then \code{bplapply} is called from the \code{BiocParallel} package;
if "future.apply" is used, then \code{future_lapply} is called from the \code{future.apply} package;
if "pbmcapply" is used, then \code{pbmclapply} is called from the \code{pbmcapply} package.
}
}
\examples{
NULL
}
