% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{fitPAModel}
\alias{fitPAModel}
\title{Fit a marginal model for power analysis}
\usage{
fitPAModel(
  df,
  model_formula,
  idx,
  method = c("nb", "poisson", "gaussian", "pseudoBulkLinear"),
  snpid,
  indiv_colname
)
}
\arguments{
\item{df}{a data frame object contains the design matrix.}

\item{model_formula}{a stats::formula object contains the model formula for power analysis.}

\item{idx}{a numeric value recording the serial number of the simulation}

\item{method}{a character object specifying the method that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{snpid}{a character object contains snpid.}

\item{indiv_colname}{a string scalar of the sample ID variable in cell covariate
of \code{marginal_list[[geneid]]$frame}. The default is "indiv".}
}
\value{
a fitted stats::model object.
}
\description{
Fit a marginal model for power analysis
}
\examples{
NULL
}
