% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMarginal.R
\name{modifyMarginalModels}
\alias{modifyMarginalModels}
\title{Modify marginal models}
\usage{
modifyMarginalModels(
  marginal_list,
  eqtlgeno_list,
  features,
  celltype,
  neg_ctrl = FALSE,
  eqtl_loc = NULL,
  eqtl_scale = eqtl_loc,
  cellstate_colname = "cell_type",
  snp_colname = "snp_id",
  verbose = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{marginal_list}{A list of marginal model objects.}

\item{eqtlgeno_list}{A list of eqtl genotypes.}

\item{features}{A scalar or vector of features (ie. genes) to apply the modifications.}

\item{celltype}{A string to specify the cell type.}

\item{neg_ctrl}{A logical value for whether to set negative control eQTL.
If neg_ctrl = TRUE, eqtl_loc option can still be used, while eqtl_scale
will have no impact. This option sets the conditional means to be the
same across genotypes (0, 1, 2).}

\item{eqtl_loc}{A non-negative numeric scalar or vector  used as multiplicative factor to
increase or decrease the celltype eQTL mean effect size at genotype 1. The
default is eqtl_loc = NULL, where no effect size changes are made.}

\item{eqtl_scale}{A non-negative numeric scalar or vector used as multiplicative factor
to adjust the slope of celltype eQTL effect size. If NULL then no effect
size changes are made.  The default is same as \code{eqtl_loc}.}

\item{cellstate_colname}{A string for cell state variable name.}

\item{snp_colname}{A string for SNP id variable name.}

\item{verbose}{A logical value for whether to output messages related to
modified parameters. The default is TRUE.}

\item{debug}{A logical for whether to output a \code{mod_list} list in addition
to \code{marginal_list}.}

\item{...}{Additional options.}
}
\value{
A list of marginal models similar to \code{marginal_list} input.
When \code{debug} = TRUE, the output have a \code{mod_list} list containing
intermediate objects in addition to \code{marginal_list}.
}
\description{
Iterate through genes and modify marginal model parameters based on user inputs
for celltype-specific eQTLs.
}
\examples{
NULL
}
