% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMarginal.R
\name{modifyModelPara}
\alias{modifyModelPara}
\title{Modify parameters of a glmmTMB model object}
\usage{
modifyModelPara(
  model_obj,
  eqtlgeno,
  celltype,
  neg_ctrl = FALSE,
  eqtl_loc = NULL,
  eqtl_scale = eqtl_loc,
  direction = NULL,
  cellstate_colname = "cell_type",
  snp_colname = "snp_id",
  verbose = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{model_obj}{A marginal model object for a gene.}

\item{eqtlgeno}{A dataframe of eqtl genotypes for a gene.}

\item{celltype}{A string to specify the cell type.}

\item{neg_ctrl}{A logical value for whether to set negative control eQTL.
If neg_ctrl = TRUE, eqtl_loc option can still be used, while eqtl_scale
will have no impact. This option sets the conditional means to be the
same across genotypes (0, 1, 2).}

\item{eqtl_loc}{A non-negative numeric value used as multiplicative factor to
increase or decrease the celltype eQTL mean effect size at genotype 1. The
default is eqtl_loc = NULL, where no effect size changes are made.}

\item{eqtl_scale}{A non-negative numeric value used as multiplicative factor
to adjust the slope of celltype eQTL effect size. If NULL then no effect
size changes are made.}

\item{direction}{A logical value to determine whether the eQTL slope is in
nominal or reverse direction.}

\item{cellstate_colname}{A string for cell state variable name.}

\item{snp_colname}{A string for SNP id variable name.}

\item{verbose}{A logical value for whether to output messages related to
modified parameters. The default is TRUE.}

\item{debug}{A logical value for whether to output intermediate objects used
for debugging purposes. The default is FALSE.}

\item{...}{Additional options.}
}
\value{
A list of dataframe of coefficients, model objects, and optional .
}
\description{
Modify parameters of a glmmTMB model object
}
\examples{
NULL
}
