% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSCE.R
\name{normalizeSCE}
\alias{normalizeSCE}
\title{Normalize data in a SingleCellExperiment object}
\usage{
normalizeSCE(
  sce,
  method = c("log1p", "seurat", "scran", "none"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{method}{A string value specifying the normalization method. Must be one
of 'log1p', seurat', 'scran', or 'none'. The default is 'log1p'.
See Normalization details.}

\item{overwrite}{A logical value on whether to overwrite the "logcounts" slot in
input sce object. Must be set to \code{overwrite = TRUE} to overwrite. The
default is FALSE.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A SingleCellExperiment object
}
\description{
Normalizes the data in a SingleCellExperiment object with specified method
and returns a SingleCellExperiment object.
}
\details{
\subsection{Normalization options}{

If 'seurat' is used, the default options from \code{NormalizeData} from the
\code{Seurat} package is used. Seurat v4 is currently supported;
If 'scran' is used, the default options from \code{computePooledFactors},
followed by \code{logNormCounts} from the \code{scuttle} package is used.
If 'log1p' is used, the \code{log1p} function from base R is used. In general,
the computational time from fast to slow is 'log1p', 'seurat', and 'scran'.
}
}
\examples{
data("example_sce")
example_sce <- normalizeSCE(example_sce, method = "log1p", overwrite = TRUE)
}
