% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCellProportion.R
\name{plotCellProp}
\alias{plotCellProp}
\title{Visualize the cell type proportions across individuals}
\usage{
plotCellProp(
  col_data,
  title = NULL,
  color_vec = NULL,
  cellstate_colname = "cell_type",
  indiv_colname = "indiv",
  width = 1,
  linewidth = 0.01
)
}
\arguments{
\item{col_data}{a data frame with two columns: individual ids with
column name \code{indiv_colname}, and cell types with column
name \code{cellstate_colname}.}

\item{title}{a string scalar specifiying the title of the output barplot.}

\item{color_vec}{a named vector of color hex codes named by the ordered
cell type names. If NULL, the cell types contained in \code{col_data} will
be ordered by alphabetical order with R default color schemes.}

\item{cellstate_colname}{a string scalar to specify the cell type variable
in \code{col_data}}

\item{indiv_colname}{a string scalar to specify the individual id variable
in \code{col_data}.}

\item{width}{an numeric scalar specifying the width of each bar
in the output barplot.}

\item{linewidth}{an numeric scalar specifying the line width of each bar
in the output barplot.}
}
\value{
outputs a ggplot2 object
}
\description{
Function visiualizes the simulated output of the cell type proportion modelling function
}
\examples{
NULL
}
