% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPbulkGeno.R
\name{plotPbulkGeno}
\alias{plotPbulkGeno}
\title{Plot pseudobulk expression vs. genotype}
\usage{
plotPbulkGeno(
  res_df,
  celltype_colname,
  snp_colname,
  show_boxplot = TRUE,
  show_meanstat = TRUE,
  show_snp = FALSE,
  ...
)
}
\arguments{
\item{res_df}{A dataframe}

\item{celltype_colname}{A string scalar for column name of cell type. The
default is "cell_type".}

\item{snp_colname}{A string scalar for column name of the SNP variable. The
default is "snp_id".}

\item{show_boxplot}{A logical scalar for whether to overlay boxplots.}

\item{show_meanstat}{A logical scalar for whether to display the mean marked
by a red dot.}

\item{show_snp}{A logical scalar for whether to display selected SNP for each
y-axis facet.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A ggplot object.
}
\description{
Creates a faceted violin plot of pseudobulk expression values across genotype
groups, with optional boxplots and mean summary points.
}
\examples{
NULL
}
