% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{powerAnalysis}
\alias{powerAnalysis}
\title{Perform a power analysis}
\usage{
powerAnalysis(
  marginal_list,
  marginal_model,
  refit = FALSE,
  refit_formula = NULL,
  geneid,
  snpid,
  type_specific,
  type_vector,
  method = c("nb", "poisson", "gaussian", "pseudoBulkLinear"),
  nindivs = NULL,
  ncells = NULL,
  nPool = NULL,
  nIndivPerPool = NULL,
  nCellPerPool = NULL,
  alpha = 0.05,
  nsims = 100,
  ncores = 1
)
}
\arguments{
\item{marginal_list}{the output of function fitMarginalPop().}

\item{marginal_model}{a character showing the model types of the full marginal model.}

\item{refit}{a boolean parameter controlling whether user want to refit the marginal full model.}

\item{refit_formula}{the formula used to refit the marginal full model.}

\item{geneid}{a character object contains geneid.}

\item{snpid}{a character object contains snpid.}

\item{type_specific}{a character object contains the name of the covariate that the analysis is specific on.}

\item{type_vector}{a vector object contains the specified type/level names of the covariate.}

\item{method}{a character object specifying the method that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}

\item{nPool}{a vector of numeric values showing how many pools of sequencing has been performed.}

\item{nIndivPerPool}{a numerical value showing how many individuals are sequenced in one pool.}

\item{nCellPerPool}{a vector of numeric values showing how many cells are sequenced in one pool.}

\item{alpha}{the p value threshold for rejecting the H0 hypothesis.}

\item{nsims}{number of simulations for calculating the power. This parameter will affect the resolution of the power value.}

\item{ncores}{number of CPU cores user wants to use.}
}
\value{
a list of named features, each containing a list with the following items:
\describe{
\item{\code{intercept}}{the intercept for the genotype effect in the specified type/level.}
\item{\code{slope}}{the slope for the genotype effect in the specified type/level.}
\item{\code{power}}{a data frame contains power values in different parameter settings.}
\item{\code{data}}{a data frame contains both the H1 and H0 genotype effect estimates in different parameter settings and simulation times.}
}
}
\description{
Perform a power analysis
}
\examples{
NULL
}
