% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{powerCICalculation}
\alias{powerCICalculation}
\title{Calculate a bootstrap confidence interval for each power}
\usage{
powerCICalculation(
  res,
  celltype_vector,
  nindivs,
  ncells,
  snp_number = 10,
  gene_number = 800,
  alpha = 0.05,
  nsim = 1000,
  conf = 0.05
)
}
\arguments{
\item{res}{the output of function powerAnalysis()}

\item{celltype_vector}{a vector object specifies the cell type that will be tested}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}

\item{snp_number}{the number of SNPs for multiple testing correction.}

\item{gene_number}{the number of genes for multiple testing correction.}

\item{alpha}{the p value threshold for rejecting the H0 hypothesis.}

\item{nsim}{number of simulations for calculating the Bootstrap CI.}

\item{conf}{Bootstrap CI interval.}
}
\value{
a data frame contains average power with standard deviations in each parameter settings.
}
\description{
Calculate a bootstrap confidence interval for each power
}
\examples{
NULL
}
