% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCellProportion.R
\name{simuCellProportion}
\alias{simuCellProportion}
\title{Simulate cell proportions using genotype principal components and population-level covariates}
\usage{
simuCellProportion(
  sce,
  genoPC,
  new_genoPC,
  new_othercov,
  PCnum = 5L,
  cov_colnames = NULL,
  indiv_colname = "indiv",
  celltype_colname = "cell_type",
  cn_model_family = "lognormal",
  cn_meanlog = NULL,
  cn_sdlog = NULL,
  cp_model_family = "MN",
  cp_intercept = TRUE,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object.}

\item{genoPC}{a data frame of individual by genotype principal components for \code{sce} input.
The first column must be the variable for individual with same name as \code{indiv_colname}.}

\item{new_genoPC}{a data frame of individual by genotype principal components for
simulated individuals. The first column must be the variable for individual
with same name as \code{indiv_colname}, followed by "PC1", "PC2", etc.}

\item{new_othercov}{a data frame of the test data containing same additional covariates
as in colData of \code{sce}.}

\item{PCnum}{an integer scalar specifying the number of principal components used
in multinomial regression.}

\item{cov_colnames}{an optional string vector or scalar for the variable names to include
in the cell proportion model. Variables must exist in both \code{new_othercov} and
colData of \code{sce}.}

\item{indiv_colname}{a string scalar to specify the variable in \code{sce} containing
individuals.}

\item{celltype_colname}{a string scalar to specify the variable in \code{sce} containing
cell types.}

\item{cn_model_family}{a string scalar to specify the model family used for total
cell modeling. Currently only 'lognormal' from \link[MASS]{fitdistr} is supported.}

\item{cn_meanlog}{a numeric scalar for the mean parameter (on log scale) of the total
cell number model. When \code{cn_meanlog = NULL}, the parameter is estimated from
input data.}

\item{cn_sdlog}{a numeric scalar for the standard deviation parameter (on log scale)
of the total cell number model. When \code{n_sdlog = NULL}, the parameter is
estimated from input data.}

\item{cp_model_family}{a string scalar to specify the model family used for cell
proportion modeling. Currently only 'MN' from \link[MGLM]{dist} is supported.}

\item{cp_intercept}{a logical scalar for whether to include an intercept in the
cell proportion model.}

\item{...}{additional optional arguments.}
}
\value{
outputs a list with following elements:
\describe{
\item{\code{simu_cov}}{a cell-by-covariate data frame of simulated cell types and corresponding individual.}
\item{\code{cp_simu_df}}{a cell type-by-covariate data frame summarizing the simulate cell proportions, total cell numbers, and cells per cell types.}
\item{\code{cp_modelfit}}{a fitted model object for the cell proportion model.}
\item{\code{cn_modelfit}}{a fitted model object for the cell number model.}
}
}
\description{
Function fits a multinomial regression model using the input genotype principal
components (PCs) and other population-level covariates as training data and outputs
simulated cell proportions.
}
\examples{
NULL
}
