% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors.gamlss}
\alias{calcParaVectors.gamlss}
\title{A calcParaVectors Method for gamlss Objects}
\usage{
\method{calcParaVectors}{gamlss}(fit, family_use, new_covariate, total_cells, data, ...)
}
\arguments{
\item{fit}{add later}

\item{family_use}{add later}

\item{new_covariate}{add later}

\item{total_cells}{add later}

\item{data}{add later}

\item{...}{Additional arguments passed to calcParaVectors S3 method functions.}
}
\description{
A calcParaVectors Method for gamlss Objects
}
