% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors.glmmTMB}
\alias{calcParaVectors.glmmTMB}
\title{A calcParaVectors Method for glmmTMB Objects}
\usage{
\method{calcParaVectors}{glmmTMB}(fit, family_use, new_covariate, total_cells, ...)
}
\arguments{
\item{fit}{add later}

\item{family_use}{add later}

\item{new_covariate}{add later}

\item{total_cells}{add later}

\item{...}{add later}
}
\description{
A calcParaVectors Method for glmmTMB Objects
}
