% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors}
\alias{calcParaVectors}
\title{Generic function to compute model parameter vectors}
\usage{
calcParaVectors(fit, family_use, new_covariate, total_cells, data, ...)
}
\arguments{
\item{fit}{a fitted object in the marginal_list.}

\item{family_use}{a string scalar or vector of marginal distribution used.}

\item{new_covariate}{a cell-by-covariate data frame obtained in the list output from
\code{\link{constructDataPop}}. It must have a corr_group variable.}

\item{total_cells}{a positive integer for the number of total cells to simulate.}

\item{data}{a cell-by-covariate data frame obtained in the list output from
\code{\link{constructDataPop}}. It must have a corr_group variable.
Used only in gamlss fits.}

\item{...}{additional arguments passed to calcParaVectors S3 method functions.}
}
\description{
A S3 generic function for computing model parameters for with or without new
covariate of a feature
}
