% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParaPop.R
\name{calcParaVectors}
\alias{calcParaVectors}
\title{Generic function to compute model parameter vectors}
\usage{
calcParaVectors(fit, family_use, new_covariate, total_cells, data, ...)
}
\arguments{
\item{fit}{add later}

\item{family_use}{add later}

\item{new_covariate}{add later}

\item{total_cells}{add later}

\item{data}{add later}

\item{...}{add later}
}
\description{
A S3 generic function for computing model parameters for with or without new
covariate of a feature
}
