% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkVectorContain}
\alias{checkVectorContain}
\title{Check membership of first vector compared to other vectors}
\usage{
checkVectorContain(..., ignore_dups = FALSE)
}
\arguments{
\item{...}{arguments of two or more vectors.}

\item{ignore_dups}{logical scalar to disregard duplicate elements during comparison.
Default is FALSE.}
}
\value{
a logical scalar
}
\description{
This is an internal helper function to check if elements of the first vector are
subsets of one or more other vectors. Ordering of elements is ignored.
}
\examples{
vec1 <- c("cherry", "apple", "cherry")
vec2 <- c("apple", "cherry", "banana", "cherry")
vec3 <- c("banana", "kiwi", "apple", "cherry")
vec4 <- c("banana", "apple", "apple", "cherry")

checkVectorContain(vec1, vec2, vec3, ignore_dups = TRUE)   # returns TRUE
checkVectorContain(vec1, vec2, vec4, ignore_dups = FALSE)  # returns FALSE
}
