% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMarginalPop.R
\name{constructDesignMatrix}
\alias{constructDesignMatrix}
\title{Construct a Design Matrix Dataframe}
\usage{
constructDesignMatrix(
  response_vec,
  cellcov_df,
  eqtlgeno_df,
  interact_colnames = NULL,
  loc_colname = "POS",
  snp_colname = "snp_id",
  cellstate_colname = "cell_type",
  indiv_colname = "indiv",
  filter_snps = TRUE,
  snpvar_thres = 0
)
}
\arguments{
\item{response_vec}{a vector of values for the response variable.}

\item{cellcov_df}{a cell-by-covariates dataframe containing the covariates
(explanatory variables) for all cells.}

\item{eqtlgeno_df}{a SNP-by-sample genotype dataframe containing a feature's
eQTL annotations and SNP genotypes (explanatory variables) for all samples
(ie. individuals).}

\item{interact_colnames}{a string scalar or vector for the variable names that
have first-order interaction with SNP genotypes.}

\item{loc_colname}{a string scalar for column name of SNP position variable.}

\item{snp_colname}{a string scalar for column name of SNP id variable.}

\item{cellstate_colname}{a string scalar for column name of cell state (ie. cell type).}

\item{indiv_colname}{a string scalar for column name of individuals (samples).}

\item{filter_snps}{a logical scalar for whether to filter out low-variance or
unique SNP covariates prior to fitting the model.}

\item{snpvar_thres}{a numeric scalar (between 0 to 1) used to filter out SNPs
whose variance of genotypes across samples are below this threshold.
Used together with \code{filter_snps} option.}
}
\value{
a list containing the following:
\describe{
\item{\code{dmat_df}}{a dataframe of the design matrix containing all covariates
(both cell covariates and eQTL genotype covariates) for a given feature.}
\item{\code{snp_cov}}{a string scalar or vector of SNP ids in the design matrix.}
}
}
\description{
A helper function that constructs a design matrix using a given feature's
expression vector for every cell, eQTL genotype dataframe, and cell
covariate dataframe.
}
\examples{
NULL
}
