% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSCE.R
\name{constructSCE}
\alias{constructSCE}
\title{Construct a SingleCellExperiment object with specified covariates.}
\usage{
constructSCE(
  data_obj,
  cellcov_df = NULL,
  featcov_df = NULL,
  cellcov_names,
  sampid_vec,
  overlap_features = NULL,
  assay_name = "RNA",
  slot_name = "counts",
  sce_name = slot_name,
  indiv_colname = "indiv",
  cellstate_colname = "cell_type",
  factor_colnames = NULL,
  cellcov_renames = cellcov_names
)
}
\arguments{
\item{data_obj}{a Seurat, SingleCellExperiment, or expression matrix (gene-by-cell)
used to construct the new SingleCellExperiment object.}

\item{cellcov_df}{a dataframe containing covariates for all cells. Row names must
have cell names that match \code{data_obj}. Only used when \code{data_obj}
is a matrix object.}

\item{featcov_df}{a dataframe containing covariates for all features (ie genes).
Row names must have feature names that match \code{data_obj}. Only used when
\code{data_obj} is a matrix object.}

\item{cellcov_names}{a string vector of all columns to extract from cell covariates.}

\item{sampid_vec}{an optional string vector of sample ids (individuals). Default
is NULL. If provided, it must match sample ids found in \code{data_obj}.}

\item{overlap_features}{an optional string vector for features to filter \code{data_obj}.
Default is NULL. If provided, all features must be present in the \code{data_obj}.}

\item{assay_name}{a string scalar of name of the Seurat object slot of interest.
Only used if \code{data_obj} is a Seurat object.}

\item{slot_name}{a string scalar for the type of assay data (either 'logcounts'
or 'counts'). Default is 'counts'. Only used if \code{data_obj} is a Seurat or
SingleCellExperiment object.}

\item{sce_name}{a string scalar for the type of assay data for output SingleCellExperiment.}

\item{indiv_colname}{a string scalar for the sample id colname. Must be present
in \code{cellcov_df} or \code{data_obj}. Default is 'indiv'.}

\item{cellstate_colname}{a string scalar for the cell state colname. Must be present
in \code{cellcov_df} or \code{data_obj}. Default is 'cell_type'.}

\item{factor_colnames}{an optional string vector or scalar for the columns to
coerce into factor variables. Default is NULL. If provided, it must be present
in \code{cellcov_df} or \code{data_obj}.}

\item{cellcov_renames}{an optional string vector to rename the cell covariates.
Must be same length as \code{cellcov_names}. Default is same as \code{cellcov_names}.}
}
\value{
a SingleCellExperiment object
}
\description{
Construct a SingleCellExperiment object with specified covariates.
}
\examples{
NULL
}
