% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtlgeno_data.R
\docType{data}
\name{example_eqtlgeno}
\alias{example_eqtlgeno}
\title{A dataframe containing both eQTL annotations and genotypes of individuals}
\format{
\subsection{\code{example_eqtlgeno}}{

A data frame with 270 cell-type specific eQTLs (rows), 5 eQTL relevant
annotations (cols) consisting of \code{cell_type}, \code{gene_id}, \code{snp_id},
\code{CHR}, and \code{POS}, and 20 samples' genotypes (cols).
\describe{
\item{cell_type}{Abbreviated cell type}
\item{gene_id}{Gene ID}
\item{snp_id}{SNP id of the cell-type specific eQTL}
\item{CHR}{Chromosome}
\item{POS}{Chromosome location of SNP}
}
}
}
\usage{
data("example_eqtlgeno")
}
\description{
A subset of data from the OneK1K dataset whose individual ids are anonymized,
and with simulated genotypes.
}
\keyword{datasets}
