% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{A SingleCellExperiment object containing both cell type and individual id}
\format{
\subsection{\code{example_sce}}{

A data frame with 100 genes (rows) and 17,918 cells (cols) for 20 individuals :
\describe{
\item{cell_type}{Cell type}
\item{indiv}{Individual id}
}
}
}
\usage{
data("example_sce")
}
\description{
A subset of the scRNA-seq data from the OneK1K cohort which has been anonymized.
}
\keyword{datasets}
