% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCopulaPop.R
\name{fitCopulaPop}
\alias{fitCopulaPop}
\title{Fits copula for input}
\usage{
fitCopulaPop(
  sce,
  assay_use,
  input_data,
  marginal_list,
  family_use,
  copula = "gaussian",
  DT = TRUE,
  pseudo_obs = FALSE,
  epsilon = 1e-06,
  family_set = c("gaussian", "indep"),
  important_feature = "all",
  n_cores,
  parallelization = "mcmapply",
  BPPARAM = NULL
)
}
\arguments{
\item{sce}{add later}

\item{assay_use}{add later}

\item{input_data}{add later}

\item{marginal_list}{add later}

\item{family_use}{add later}

\item{copula}{add later}

\item{DT}{add later}

\item{pseudo_obs}{add later}

\item{epsilon}{add later}

\item{family_set}{add later}

\item{important_feature}{add later}

\item{n_cores}{add later}

\item{parallelization}{add later}

\item{BPPARAM}{add later}
}
\value{
A list
}
\description{
Fits copula for input
}
\examples{
NULL
}
