% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMarginalPop.R
\name{fitMarginalPop}
\alias{fitMarginalPop}
\title{Fit marginal models for every feature}
\usage{
fitMarginalPop(
  data_list,
  mean_formula,
  model_family = "nb",
  interact_colnames = NULL,
  parallelization = "pbmcapply",
  n_threads = 2L,
  loc_colname = "POS",
  snp_colname = "snp_id",
  cellstate_colname = "cell_type",
  indiv_colname = "indiv",
  filter_snps = TRUE,
  snpvar_thres = 0,
  force_formula = FALSE,
  data_maxsize = 1
)
}
\arguments{
\item{data_list}{a list of input data.}

\item{mean_formula}{a string scalar to specify the mean formula, including
random effects (if any) but without SNP genotypes or SNP genotype interaction effects.}

\item{model_family}{a string scalar to specify model fitting used.}

\item{interact_colnames}{a string scalar or vector for the variable names that
have first-order interaction with SNP genotypes.}

\item{parallelization}{a string scalar specifying the type of parallization
used during marginal fitting.}

\item{n_threads}{positive integer value (greater or equal to 1) to specify the
number of CPU threads used in parallelization.}

\item{loc_colname}{a string scalar for column name of SNP position variable.}

\item{snp_colname}{a string scalar for column name of SNP id variable.}

\item{cellstate_colname}{a string scalar for column name of cell state (ie. cell type).}

\item{indiv_colname}{a string scalar for column name of individuals (samples).}

\item{filter_snps}{a logical scalar for whether to filter out low-variance or
unique SNP covariates prior to fitting the model.}

\item{snpvar_thres}{a numeric scalar (between 0 to 1) used to filter out SNPs
whose variance of genotypes across samples are below this threshold.
Used together with \code{filter_snps} option.}

\item{force_formula}{a logical scalar for whether to allow interaction terms
without the covariate being a main effect in the model.}

\item{data_maxsize}{a positive numeric value used to set max data_list size
in GiB unit. Used only when \code{parallelization = 'future'}.}
}
\value{
a list of named features, each containing a list with the following items:
\describe{
\item{\code{fit}}{a \code{glmmTMB} fit object.}
\item{\code{time}}{a numeric scalar for the elapsed time to fit the given feature.}
\item{\code{snp_cov}}{a string scalar or vector of SNP ids used in the
fit for given feature.}
\item{\code{removed_cell}}{a string scalar or vector of the cell names
removed due to low-variance (NOT currently implemented).}
}
}
\description{
Fits a specified parametric model using various input parameters.
}
\examples{
NULL
}
