% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMarginal.R
\name{modifyMarginalModels}
\alias{modifyMarginalModels}
\title{Modify marginal models}
\usage{
modifyMarginalModels(
  marginal_list,
  eqtlgeno_list,
  features,
  celltype,
  neg_ctrl = FALSE,
  mean_log2fc = 0,
  eqtl_log2fc = mean_log2fc,
  eqtl_reverse = FALSE,
  mean_baseline = NULL,
  eqtl_baseline = NULL,
  mean_baseline_only = FALSE,
  eqtl_baseline_only = FALSE,
  disp_scaling = "linear",
  celltype_colname = "cell_type",
  snp_colname = "snp_id",
  verbose = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{marginal_list}{A list of marginal model objects.}

\item{eqtlgeno_list}{A list of eqtl genotypes.}

\item{features}{A scalar or vector of features (ie. genes) to apply the modifications.}

\item{celltype}{A string to specify the cell type in which to make the modification.}

\item{neg_ctrl}{A logical value for whether to set a negative control eQTL (ie. a
non-eGene).  This option sets the conditional means to be identical across
genotypes (0, 1, 2).  If \code{neg_ctrl = TRUE}, the \code{mean_log2fc} option
will still be applied if set, but eqtl_log2fc will be overidden and have
no impact.  Default is \code{FALSE}.}

\item{mean_log2fc}{A numeric scalar or vector for the log2 fold-change parameter to
increase or decrease the conditional mean at genotype 1 \eqn{\mu_{1}} in
a cell type. Default is \code{mean_log2fc = 0} (no parameters are modified
and uses estimated parameters from the fitted marginal model).}

\item{eqtl_log2fc}{A numeric scalar or vector for the log2 fold-change parameter to
increase or decrease the slope of eQTL effect in a celltype. The eQTL slope
is defined as the difference between the conditional mean at genotype 1
and genotype 0 (\eqn{\mu_{1}} - \eqn{\mu_{0}}). Default is
\code{eqtl_log2fc = mean_log2fc} (eQTL slope is scaled the same as the
conditional mean log2 fold-change).}

\item{eqtl_reverse}{A logical value to determine whether the eQTL slope trends
in the reverse direction (TRUE) or same (FALSE). Default is \code{FALSE}.}

\item{mean_baseline}{A numeric scalar or vector to specify the minimum conditional mean
at genotype 1 \eqn{\mu_{1}}.  If \code{mean_baseline_only = FALSE},
then the conditional mean will be the maximum of the fitted (estimated from
marginal model) and the \code{mean_baseline} value.  Otherwise, the
conditional mean will be set to the \code{mean_baseline} value.  Default
value is \code{NULL}.}

\item{eqtl_baseline}{A numeric scalar or vector to specify the minimum eQTL slope
between genotype 1 and 0 (\eqn{\mu_{1}} - \eqn{\mu_{0}}).  If
\code{eqtl_baseline_only = FALSE}, then the eQTL slope will be the
maximum of the slope of fitted (estimated from marginal model) and
the \code{eqtl_baseline} value.  Otherwise, the eQTL slope will be set to
the \code{eqtl_baseline} value.  Default value is \code{NULL}.}

\item{mean_baseline_only}{A logical scalar or vector to force the conditional mean (in
linear prediction) at genotype 1 \eqn{\mu_{1}}. Default is \code{FALSE}.}

\item{eqtl_baseline_only}{A logical scalar or vector to force the eQTL slope between
genotype 1 and 0 (\eqn{\mu_{1}} - \eqn{\mu_{0}}). Default is \code{FALSE}.}

\item{disp_scaling}{A string value to specify the dispersion-mean scaling for
certain parametric models. Current options are either \code{"linear"},
\code{"quadratic"}, or \code{"none"}. (NOTE: currently only applicable to
the negative binomial model.)}

\item{celltype_colname}{A string for cell type variable name.}

\item{snp_colname}{A string for SNP id variable name.}

\item{verbose}{A logical value for whether to output messages related to
modified parameters. Default is \code{TRUE}.}

\item{debug}{A logical for whether to output a \code{mod_list} list in addition
to \code{marginal_list}.}

\item{...}{Additional options.}
}
\value{
A list of marginal models similar to \code{marginal_list} input.
When \code{debug} = TRUE, the output have a \code{mod_list} list containing
intermediate objects in addition to \code{marginal_list}.
}
\description{
Modify the marginal model parameters of genes based on user inputs for
cell-type-specific eQTLs.
}
\examples{
NULL
}
