% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{powerAnalysis}
\alias{powerAnalysis}
\title{Perform a cell-type-specific power analysis on eQTL effects}
\usage{
powerAnalysis(
  marginal_list,
  marginal_model = NULL,
  refit_formula = NULL,
  geneid = NULL,
  snpid = NULL,
  cellstate_colname = "cell_type",
  cellstate_vector = NULL,
  indiv_colname = "indiv",
  method = c("nb", "poisson", "gaussian", "pseudoBulkLinear"),
  nindivs = NULL,
  ncells = NULL,
  nPool = NULL,
  nIndivPerPool = NULL,
  nCellPerPool = NULL,
  alpha = 0.05,
  nsims = 100,
  ncores = 1
)
}
\arguments{
\item{marginal_list}{the output of function fitMarginalPop().}

\item{marginal_model}{a character showing the model types of the full marginal model.}

\item{refit_formula}{the formula used to refit the marginal full model if user wants to. Default is null.}

\item{geneid}{a character object contains geneid.}

\item{snpid}{a character object contains snpid.}

\item{cellstate_colname}{a string scalar specifying the cell state variable in
\code{marginal_list[[geneid]]$frame}.
The default is "cell_type".}

\item{cellstate_vector}{a vector object specifies the cell type that will be tested}

\item{indiv_colname}{a string scalar of the sample ID variable in cell covariate
of \code{marginal_list[[geneid]]$frame}. The default is "indiv".}

\item{method}{a character object specifying the method that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}

\item{nPool}{a vector of numeric values showing how many pools of sequencing has been performed.}

\item{nIndivPerPool}{a numerical value showing how many individuals are sequenced in one pool.}

\item{nCellPerPool}{a vector of numeric values showing how many cells are sequenced in one pool.}

\item{alpha}{the p value threshold for rejecting the H0 hypothesis.}

\item{nsims}{number of simulations for calculating the power. This parameter will affect the resolution of the power value.}

\item{ncores}{number of CPU cores user wants to use.}
}
\value{
a list of named features, each containing a list with the following items:
\describe{
\item{\code{intercept}}{the intercept for the genotype effect in the specified type/level.}
\item{\code{slope}}{the slope for the genotype effect in the specified type/level.}
\item{\code{power}}{a data frame contains power values in different parameter settings.}
\item{\code{data}}{a data frame contains both the H1 and H0 genotype effect estimates in different parameter settings and simulation times.}
}
}
\description{
Perform a cell-type-specific power analysis on eQTL effects
}
\examples{
NULL
}
