% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{runPowerAnalysis}
\alias{runPowerAnalysis}
\title{The wrapper function for power analysis}
\usage{
runPowerAnalysis(
  marginal_list,
  marginal_model = "nb",
  refit_formula = NULL,
  geneid = NULL,
  snpid = NULL,
  celltype_colname = "cell_type",
  celltype_vector = NULL,
  celltype_specific_ES_list = NULL,
  indiv_colname = "indiv",
  methods = NULL,
  nindivs = NULL,
  ncells = NULL,
  nPool = NULL,
  nIndivPerPool = NULL,
  nCellPerPool = NULL,
  alpha = 0.05,
  power_nsim = 100,
  snp_number = 10,
  gene_number = 800,
  CI_nsim = 1000,
  CI_conf = 0.05,
  ncores = 2L,
  parallelization = c("pbmcapply", "future.apply", "parallel", "biocparallel"),
  BPPARAM = NULL,
  future.seed = FALSE,
  data_maxsize = 1
)
}
\arguments{
\item{marginal_list}{the output of function fitMarginalPop().}

\item{marginal_model}{a character showing the model types of the full marginal model.}

\item{refit_formula}{the formula used to refit the marginal full model if user wants to. Default is null.}

\item{geneid}{a character object contains geneid.}

\item{snpid}{a character object contains snpid.}

\item{celltype_colname}{a string scalar specifying the cell state variable in
\code{marginal_list[[geneid]]$frame}.
The default is "cell_type".}

\item{celltype_vector}{a vector object specifies the cell type that will be tested}

\item{celltype_specific_ES_list}{a list object specifies different vectors of the genotype effect size (ES) for each cell type}

\item{indiv_colname}{a string scalar of the sample ID variable in cell covariate
of \code{marginal_list[[geneid]]$frame}. The default is "indiv".}

\item{methods}{a vector of character objects specifying the methods that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}

\item{nPool}{a vector of numeric values showing how many pools of sequencing has been performed.}

\item{nIndivPerPool}{a numerical value showing how many individuals are sequenced in one pool.}

\item{nCellPerPool}{a vector of numeric values showing how many cells are sequenced in one pool.}

\item{alpha}{the p value threshold for rejecting the H0 hypothesis.}

\item{power_nsim}{a number of simulations for calculating the power. This parameter will affect the resolution of the power value.}

\item{snp_number}{the number of SNPs for multiple testing correction.}

\item{gene_number}{the number of genes for multiple testing correction.}

\item{CI_nsim}{number of simulations for calculating the Bootstrap CI.}

\item{CI_conf}{Bootstrap CI interval.}

\item{ncores}{a positive integer value (greater or equal to 1) to specify the
number of CPU cores used in parallelization. The default is 2.}

\item{parallelization}{a string scalar specifying the parallelization backend
used when simulating data. Must be one of "parallel", "future.apply",
"biocparallel", or "pbmcapply". The default value is "pbmcapply". See details.}

\item{BPPARAM}{a BiocParallelParam class object (from \code{BiocParallel} R package)
that must be specified when using \code{parallelization = "biocparallel"}. Either
\code{BiocParallel::SnowParam()} or \code{BiocParallel::MulticoreParam()}
can be used to initialize, depending on the operating system. BPPARAM is
not used in other parallelization options. The default is NULL.}

\item{future.seed}{a logical or an integer (of length one or seven), or a list
of length(X) with pre-generated random seeds that can be specified when using
\code{parallelization = "future.apply"}. See \code{future.apply::future_eapply}
documentation for more details on its usage. future.seed is not used in
other parallelization options. The default is FALSE.}

\item{data_maxsize}{a positive numeric value used to set max marginal_list size
in GiB increments. Used only when \code{parallelization = "future.apply"}.
The default is 1.}
}
\value{
a data frame contains power analysis result in different parameter settings.
}
\description{
\subsection{Parallelization options}{

If "parallel" is used then \code{mclapply} is called from the \code{parallel} package; if
"biocparallel" is used, then \code{bplapply} is called from the \code{BiocParallel} package; if
"future.apply" is used, then \code{future_lapply} is called from the \code{future.apply} package;
if "pbmcapply" is used, then \code{pbmclapply} is called from the \code{pbmcapply} package.
}
}
\examples{
NULL
}
