% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{runPowerAnalysis}
\alias{runPowerAnalysis}
\title{The wrapper function for power analysis}
\usage{
runPowerAnalysis(
  marginal_list,
  marginal_model,
  refit = FALSE,
  refit_formula = NULL,
  geneid,
  snpid,
  type_specific,
  type_vector,
  methods,
  nindivs = NULL,
  ncells = NULL,
  nPool = NULL,
  nIndivPerPool = NULL,
  nCellPerPool = NULL,
  alpha = 0.05,
  power_nsim = 100,
  snp_number = 10,
  gene_number = 800,
  CI_nsim = 1000,
  CI_conf = 0.05,
  ncores = 1
)
}
\arguments{
\item{marginal_list}{the output of function fitMarginalPop().}

\item{marginal_model}{a character showing the model types of the full marginal model.}

\item{refit}{a boolean parameter controlling whether user want to refit the marginal full model.}

\item{refit_formula}{the formula used to refit the marginal full model.}

\item{geneid}{a character object contains geneid.}

\item{snpid}{a character object contains snpid.}

\item{type_specific}{a character object contains the name of the covariate that the analysis is specific on.}

\item{type_vector}{a vector object contains the specified type/level names of the covariate.}

\item{methods}{a vector of character objects specifying the methods that will be analyzed for power.
(Options: nb,poisson,gaussian,pseudoBulkLinear).}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}

\item{nPool}{a vector of numeric values showing how many pools of sequencing has been performed.}

\item{nIndivPerPool}{a numerical value showing how many individuals are sequenced in one pool.}

\item{nCellPerPool}{a vector of numeric values showing how many cells are sequenced in one pool.}

\item{alpha}{the p value threshold for rejecting the H0 hypothesis.}

\item{power_nsim}{a number of simulations for calculating the power. This parameter will affect the resolution of the power value.}

\item{snp_number}{the number of SNPs for multiple testing correction.}

\item{gene_number}{the number of genes for multiple testing correction.}

\item{CI_nsim}{number of simulations for calculating the Bootstrap CI.}

\item{CI_conf}{Bootstrap CI interval.}

\item{ncores}{number of CPU cores user wants to use.}
}
\value{
a data frame contains power analysis result in different parameter settings.
}
\description{
The wrapper function for power analysis
}
\examples{
NULL
}
