% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesignPop-package.R
\docType{package}
\name{scDesignPop-package}
\alias{scDesignPop}
\alias{scDesignPop-package}
\title{scDesignPop: Simulator for population-scale single-cell expression RNA-sequencing data}
\description{
Simulate realistic population-scale single-cell RNA-sequencing (scRNA-seq) data for single-cell expression quantitative trait loci analysis (sc-eQTL) using input eQTL data and scRNA-seq data. We use scDesign3's modeling framework in this package.
}
\author{
\strong{Maintainer}: Chris Y. Dong \email{chris.dong@ucla.edu} (\href{https://orcid.org/0009-0006-2282-0394}{ORCID})

Authors:
\itemize{
  \item Yihui Cen \email{yihuicen@g.ucla.edu} (\href{https://orcid.org/0000-0002-5347-0888}{ORCID})
}

}
\keyword{internal}
