% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{SimulatePADesignMatrix}
\alias{SimulatePADesignMatrix}
\title{Simulate new design matrix for power analysis}
\usage{
SimulatePADesignMatrix(
  fit,
  df_sel,
  nindiv_total,
  model,
  snpid,
  nindivs,
  ncells
)
}
\arguments{
\item{fit}{a fitted stats::model object.}

\item{df_sel}{a data frame contains the new design matrix.}

\item{nindiv_total}{a vector contains the number of individuals for each genotype.}

\item{model}{a character showing the model types of the full marginal model.}

\item{snpid}{a character object containes snpid.}

\item{nindivs}{a vector of numeric values showing the numbers of individuals that user wants to simulate.}

\item{ncells}{a vector of numeric values showing the numbers of cells per each individual that user wants to simulate.}
}
\value{
a new data frame contains the design matrix with simulated response.
}
\description{
Simulate new design matrix for power analysis
}
\examples{
NULL
}
