% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{simulatePADesignMatrix}
\alias{simulatePADesignMatrix}
\title{Simulate new design matrix for power analysis}
\usage{
simulatePADesignMatrix(
  fit,
  df_sel,
  nindiv_total,
  model = c("nb", "poisson", "gaussian"),
  snpid,
  nindiv,
  ncell,
  celltype_colname,
  indiv_colname
)
}
\arguments{
\item{fit}{a fitted stats::model object.}

\item{df_sel}{a data frame contains the new design matrix.}

\item{nindiv_total}{a vector contains the number of individuals for each genotype.}

\item{model}{a character showing the model types of the full marginal model.}

\item{snpid}{a character object contains snpid.}

\item{nindiv}{a numeric value showing the number of individuals that user wants to simulate.}

\item{ncell}{a numeric value showing the number of cells per each individual that user wants to simulate.}

\item{celltype_colname}{a string scalar specifying the cell state variable in
\code{df_sel}.
The default is "cell_type".}

\item{indiv_colname}{a string scalar of the sample ID variable in cell covariate
of \code{marginal_list[[geneid]]$frame}. The default is "indiv".}
}
\value{
a new data frame contains the design matrix with simulated response.
}
\description{
Simulate new design matrix for power analysis
}
\examples{
NULL
}
