% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuNewPop.R
\name{simuNewPop}
\alias{simuNewPop}
\title{Simulate new data}
\usage{
simuNewPop(
  sce,
  assay_use = "counts",
  mean_mat,
  sigma_mat,
  zero_mat,
  quantile_mat = NULL,
  copula_list,
  n_cores,
  fastmvn = FALSE,
  family_use,
  nonnegative = TRUE,
  nonzerovar = FALSE,
  input_data,
  new_covariate,
  important_feature = "all",
  parallelization = "mcmapply",
  BPPARAM = NULL,
  filtered_gene
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{assay_use}{A string which indicates the assay you will use in the sce.
Default is 'counts'.}

\item{mean_mat}{A cell by feature matrix of the mean parameter.}

\item{sigma_mat}{A cell by feature matrix of the sigma parameter.}

\item{zero_mat}{A cell by feature matrix of the zero-inflation parameter.}

\item{quantile_mat}{A cell by feature matrix of the multivariate quantile.}

\item{copula_list}{A list of copulas for generating the multivariate quantile
matrix. If provided, the \code{quantile_mat} must be NULL.}

\item{n_cores}{An integer. The number of cores to use.}

\item{fastmvn}{An logical variable. If TRUE, the sampling of multivariate Gaussian is done
by \code{mvnfast}, otherwise by \code{mvtnorm}. Default is FALSE.}

\item{family_use}{A string of the marginal distribution. Must be one of
'poisson', 'nb', or 'gaussian'.}

\item{nonnegative}{A logical variable. If TRUE, values < 0 in the synthetic
data will be converted to 0. Default is TRUE (since the expression
matrix is nonnegative).}

\item{nonzerovar}{A logical variable. If TRUE, for any gene with zero variance,
a cell will be replaced with 1. This is designed for avoiding
potential errors, for example, PCA.}

\item{input_data}{A input count matrix.}

\item{new_covariate}{A data.frame which contains covariates of targeted simulated
data from  \code{\link{constructDataPop}}.}

\item{important_feature}{important_feature A string or vector which indicates whether a gene will be used in correlation estimation or not. If this is a string, then
this string must be either "all" (using all genes) or "auto", which indicates that the genes will be automatically selected based on the proportion of zero expression across cells
for each gene. Gene with zero proportion greater than 0.8 will be excluded form gene-gene correlation estimation. If this is a vector, then this should
be a logical vector with length equal to the number of genes in \code{sce}. \code{TRUE} in the logical vector means the corresponding gene will be included in
gene-gene correlation estimation and \code{FALSE} in the logical vector means the corresponding gene will be excluded from the gene-gene correlation estimation.
The default value for is "all".}

\item{parallelization}{A string indicating the specific parallelization function to use.
Must be one of 'mcmapply', 'bpmapply', or 'pbmcmapply', which corresponds to the parallelization function in the package
\code{parallel},\code{BiocParallel}, and \code{pbmcapply} respectively. The default value is 'mcmapply'.}

\item{BPPARAM}{A \code{MulticoreParam} object or NULL. When the parameter parallelization = 'mcmapply' or 'pbmcmapply',
this parameter must be NULL. When the parameter parallelization = 'bpmapply',  this parameter must be one of the
\code{MulticoreParam} object offered by the package 'BiocParallel. The default value is NULL.}

\item{filtered_gene}{A vector or NULL which contains genes that are excluded in the marginal and copula fitting}
}
\value{
A feature by cell matrix of the new simulated count (expression) matrix or sparse matrix.
}
\description{
\code{simuNewPop} generates new simulated data based on fitted marginal and copula models.
This function is adapted from simu_new function in scDesign3 v0.99.7
}
\details{
The function takes the new covariate (if use) from \code{\link{constructDataPop}},
parameter matrices from \code{\link{extractParaPop}} and multivariate Unifs
from \code{\link{fitCopulaPop}}.
}
\examples{
NULL
}
