% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{visualizePowerResult}
\alias{visualizePowerResult}
\title{Visualize the power analysis result}
\usage{
visualizePowerResult(
  power_result,
  celltypes,
  x_axis = "nindiv",
  y_axis = "ncell",
  col_group = "method",
  geneid,
  snpid
)
}
\arguments{
\item{power_result}{a data frame contains power analysis result in different parameter settings.}

\item{celltypes}{a vector of cell types that will be selected for visualization.}

\item{x_axis}{a character that specifies the x axis. Default is the number of individuals.}

\item{y_axis}{a character that specifies the y axis. Default is the number of cells per individual.}

\item{col_group}{a character that specifies the color groups. Default is the eQTL model.}

\item{geneid}{a character object contains geneid to be part of the plot title.}

\item{snpid}{a character object contains snpid to be part of the plot title.}
}
\value{
a ggplot object
}
\description{
Visualize the power analysis result
}
\examples{
NULL
}
